fix images in comment edit staying as b64 instead of uploading
