+++ assets/src/controllers/frontend/CommentEditor.js
@@ -1,5 +1,6 @@
 import { Controller } from '@hotwired/stimulus';
 import { QuillEditor } from '../../components/QuillEditor';
+import { uploadEmbeddedImages } from '../../services/media';
 
 export class CommentEditor extends Controller {
   static targets = ['editButton', 'editorContainer'];
@@ -61,7 +62,8 @@ export class CommentEditor extends Controller {
   }
 
   async save() {
-    const res = await fetch(this.updateUrlValue, { method: 'POST', body: this.editor.root.innerHTML });
+    const content = await uploadEmbeddedImages(this.editor.getSemanticHTML());
+    const res = await fetch(this.updateUrlValue, { method: 'POST', body: content });
     const newContent = await res.text();
 
     const richText = this.element.querySelector('.rich-text');
